+++ docs/release_notes.rst
@@ -9,6 +9,7 @@ Release notes for **Shopit**.
 =====
 
 * Add support for *Django 1.11* & *DjangoSHOP 0.12.x*.
+* Add setting ``SHOPIT_DEFAULT_PRODUCT_ORDER`` to control default ordering of products.
 * Add ability to override ``ProductSerializer`` fields through the ``fields`` GET property.
 * Add ``attribute_choices`` to product serializer fields.
 * Add ``template`` field to Flag model, adn a ``SHOPIT_FLAG_TEMPLATES`` setting.